refactor : DijkstraPathFinder 클래스를 application 레이어로 이관
