+++ src/main/java/nextstep/subway/applicaion/DijkstraPathFinder.java
@@ -1,7 +1,10 @@
-package nextstep.subway.domain;
+package nextstep.subway.applicaion;
 
 import java.util.Collection;
 import java.util.Optional;
+import nextstep.subway.domain.PathFinder;
+import nextstep.subway.domain.Section;
+import nextstep.subway.domain.Station;
 import nextstep.subway.domain.vo.Path;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
